requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# Document parsing
PyMuPDF==1.25.3
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
tabulate==0.9.0

# RAG pipeline
chromadb==0.6.3
sentence-transformers==3.4.1

# LLM
openai==1.59.3
